+++ src/services/Fields.php
@@ -1346,6 +1346,9 @@ class Fields extends Component
         $oldSettings = $fieldRecord->getOldAttribute('settings');
         $oldField = !$isNewField ? $this->getFieldById($fieldRecord->id) : null;
 
+        // For control panel save requests, make sure we have all the custom data already saved on the object.
+        $field = $this->_savingFields[$fieldUid] ?? null;
+
         // Fire a 'beforeApplyFieldSave' event
         if ($this->hasEventHandlers(self::EVENT_BEFORE_APPLY_FIELD_SAVE)) {
             $this->trigger(self::EVENT_BEFORE_APPLY_FIELD_SAVE, new ApplyFieldSaveEvent([
@@ -1401,8 +1404,11 @@ class Fields extends Component
         // Tell the current CustomFieldBehavior class about the field
         CustomFieldBehavior::$fieldHandles[$fieldRecord->handle] = true;
 
-        // Now get the field with its new settings intact
-        $field = $this->getFieldById($fieldRecord->id);
+        // Now get the field, if it's not a field save request
+        $field ??= $this->getFieldById($fieldRecord->id);
+        if ($isNewField) {
+            $field->id = $fieldRecord->id;
+        }
 
         if (!$isNewField) {
             // Set the old field handle and settings on the model in case the field type needs to do something with it